Wrapping up
Changed Animation montages and related collision functions for the Enemy's Left and Right CombatCollisions.
There are now separate events and functions to Activate and Deactivate collision for both of them. 

Also created a new project and copied only the content used in the game. The size as a result was only 3.5GB compared to the original project's 31.5GB.

Also packaged and tested the project. No problems found.