Fixed Loading and Saving Bugs
Fixed a bug where the game would crash if 'Load' in the pause menu is pressed right in the beginning of the game.

Another bug was fixed when the equipped weapon would be removed when the player transitioned to a new level.

Will have to implement a checkpoint system to help the player spawn at the last saved location (i.e. use the saved location) if they wish to load again while in the same map.