Added Shield C++ Class (Based on Item)
For blocking Enemy attacks. Uses the RMB.
Added basic functions and data members. 
Equip functionality done successfully. 
When equipped, it gets attached to a Socket on the left hand of the Player.
Imported a few Shield meshes also.
Will be changing the ArmedIdleWalkRun blendspace to a one with the Player holding a shield and a weapon, instead of only a weapon.