Item Assets added + Main Character Mesh Changed,  Lessened commit size
Current Character is the UE4 mannequin. 


Improvements to Blendspaces and Main Anim Blueprint

Animations are smoother. Added a new enum class "PlayerState". This will be used to switch between multiple states in the Anim BP more easily.


Commit wasn't being pushed to the remote because of the size of the Materials. So the Materials folder will be ignored.